+++ src/api/Light.h
@@ -11,8 +11,8 @@ class Light
 public:
     virtual void SetEnabled(const bool& State) = 0;
     virtual const bool & GetEnabled() const = 0;
-    virtual const glm::vec4 GetColor() const = 0;
-    virtual void SetColor(const glm::vec4& color) = 0;
+    virtual const glm::ivec4 GetColor() const = 0;
+    virtual void SetColor(const glm::ivec4& color) = 0;
 };
 
 #endif //LII_LIGHT_H
